+++ backend/ServiceRateBackend/src/main/java/at/hcw/serviceratebackend/dto/CreateServiceRequest.java
@@ -1,14 +1,15 @@
 package at.hcw.serviceratebackend.dto;
 
-import java.util.UUID;
+import java.math.BigDecimal;
 import java.util.List;
+import java.util.UUID;
 
 public record CreateServiceRequest(
         UUID providerId, // Wer bietet das an? (Die ID, die du gerade in der Datenbank gesehen hast!)
         String title,
         String description,
         String category,
-        Double price,
+        BigDecimal price,
         Double estimatedHours,
         String imageUrl,
         List<String> imageUrls,
