+++ backend/ServiceRateBackend/src/main/java/at/hcw/serviceratebackend/dto/ServiceOfferingResponse.java
@@ -1,5 +1,6 @@
 package at.hcw.serviceratebackend.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import java.util.List;
 
@@ -11,7 +12,7 @@ public record ServiceOfferingResponse(
         String title,
         String description,
         String category,
-        Double price,
+        BigDecimal price,
         Double estimatedHours,
         String imageUrl,
         List<String> imageUrls,
